requirements.txt
streamlit>=1.28.0
openai>=1.3.0
python-dotenv>=1.0.0
st-copy-to-clipboard>=0.1.0